video_classificaion_model/requirements.txt
scikit-learn>=1.0.2
tqdm
pyyaml
opencv-python
matplotlib
tensorboard
Pillow